@@ -1,149 +1,146 @@
-import 'package:memory_game/models/TileModel.dart';
+import 'package:memory_game/models/tile_model.dart';
 
 String selectedTile = "";
-int selectedIndex ;
+int? selectedIndex;
 bool selected = true;
 int points = 0;
 
-List<TileModel> myPairs = new List<TileModel>();
-List<bool> clicked = new List<bool>();
+List<TileModel> myPairs = [];
+List<bool> clicked = [];
 
-List<bool> getClicked(){
-
-  List<bool> yoClicked = new List<bool>();
-  List<TileModel> myairs = new List<TileModel>();
+List<bool> getClicked() {
+  List<bool> yoClicked = [];
+  List<TileModel> myairs = [];
   myairs = getPairs();
-  for(int i=0;i<myairs.length;i++){
+  for (int i = 0; i < myairs.length; i++) {
     yoClicked[i] = false;
   }
 
   return yoClicked;
 }
 
-List<TileModel>  getPairs(){
+List<TileModel> getPairs() {
+  List<TileModel> pairs = [];
 
-  List<TileModel> pairs = new List<TileModel>();
-
-  TileModel tileModel = new TileModel();
+  TileModel tileModel = TileModel();
 
   //1
   tileModel.setImageAssetPath("assets/fox.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //2
   tileModel.setImageAssetPath("assets/hippo.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //3
   tileModel.setImageAssetPath("assets/horse.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //4
   tileModel.setImageAssetPath("assets/monkey.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
   //5
   tileModel.setImageAssetPath("assets/panda.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //6
   tileModel.setImageAssetPath("assets/parrot.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //7
   tileModel.setImageAssetPath("assets/rabbit.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //8
   tileModel.setImageAssetPath("assets/zoo.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   return pairs;
 }
 
-List<TileModel>  getQuestionPairs(){
+List<TileModel> getQuestionPairs() {
+  List<TileModel> pairs = [];
 
-  List<TileModel> pairs = new List<TileModel>();
-
-  TileModel tileModel = new TileModel();
+  TileModel tileModel = TileModel();
 
   //1
   tileModel.setImageAssetPath("assets/question.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //2
   tileModel.setImageAssetPath("assets/question.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //3
   tileModel.setImageAssetPath("assets/question.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //4
   tileModel.setImageAssetPath("assets/question.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
   //5
   tileModel.setImageAssetPath("assets/question.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //6
   tileModel.setImageAssetPath("assets/question.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //7
   tileModel.setImageAssetPath("assets/question.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   //8
   tileModel.setImageAssetPath("assets/question.png");
   tileModel.setIsSelected(false);
   pairs.add(tileModel);
   pairs.add(tileModel);
-  tileModel = new TileModel();
+  tileModel = TileModel();
 
   return pairs;
 }
